\name{ant}
\alias{ant}
\docType{data}
\title{
Ant species incidence frequencies data
}
\description{
Species incidence frequencies for ant samples from five elevations in northeastern Costa Rica (Longino & Colwell 2011)
}
\usage{data(ant)}
\format{
  For incidence data, the first entry must be the total number of sampling units, followed by the species incidence frequencies. The format is: \cr
List of 5 \cr
 $ h50m  : num [1:228] 599 1 1 1 1 1 1 1 1 1 ... \cr
 $ h500m : num [1:242] 230 1 1 1 1 1 1 1 1 1 ... \cr
 $ h1070m: num [1:123] 150 1 1 1 1 1 1 1 1 1 ... \cr
 $ h1500m: num [1:57] 200 1 1 1 1 1 1 1 1 1 ... \cr
 $ h2000m: num [1:15] 200 1 2 2 3 4 8 8 13 15 ... \cr
 }
\references{
Longino, J.T. & Colwell, R.K. (2011) Density compensation, species composition, and richness of ants on a neotropical elevational gradient. Ecosphere, 2, art29.
}

\keyword{datasets}
