% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invChat.R
\name{as.abucount}
\alias{as.abucount}
\title{Transform abundance raw data to abundance row-sum counts (iNEXT input format)}
\usage{
as.abucount(x)
}
\arguments{
\item{x}{a \code{data.frame} or \code{matirx} of species by sites matrix.}
}
\value{
a \code{vector} of species abundance row-sum counts.
}
\description{
\code{as.abucount}: transform abundance raw data (a species by sites matrix) to abundance rwo-sum counts data (iNEXT input format).
}
\examples{
data(ciliates)
lapply(ciliates, as.abucount)

}

