% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXT.r
\name{AsyD}
\alias{AsyD}
\title{Asymptotic diversity q profile}
\usage{
AsyD(x, q = seq(0, 2, 0.2), datatype = "abundance", nboot = 50, conf = 0.95)
}
\arguments{
\item{x}{a matrix/data.frame (species by sites), or list of species abundances or incidence frequencies. If \code{datatype = "incidence_freq"}, then the first entry of the input data must be total number of sampling units in each column or list.}

\item{q}{a nonnegative value or sequence specifying the diversity order. Default is seq(0, 2, by = 0.2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species-by-site incidence frequencies data (\code{datatype = "incidence_freq"}). Default is "abundance".}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}
}
\value{
a table of diversity q profile by 'Estimated' and 'Empirical'
}
\description{
\code{AsyD} The estimated and empirical diversity of order q
}
\examples{
## example for abundance based data (list of vector)
# abundance data
data(spider)
out1 <- AsyD(spider, datatype = "abundance")
out1

## Not run:
## example for incidence frequencies based data (list of data.frame)
data(ant)
out2 <- AsyD(ant, datatype = "incidence_freq")
out2

## End(Not run)

}
\references{
Chao,A. and Jost,L.(2015).Estimating diversity and entropy profiles via discovery rates of new species.
}
