\name{Chat.Ind}
\alias{Chat.Ind}
\title{Abundance-based sample coverage}
\usage{
  Chat.Ind(x, m)
}
\arguments{
  \item{x}{a vector of species abundances}

  \item{m}{a integer vector of rarefaction/extrapolation
  sample size}
}
\value{
  a vector of estimated sample coverage function
}
\description{
  \code{Chat.Ind} Estimation of abundance-based sample
  coverage function
}

