\name{conf.reg}
\alias{conf.reg}
\title{Draw confidence band plot}
\usage{
  conf.reg(x, LCL, UCL, ...)
}
\arguments{
  \item{x}{a vector of estimator}

  \item{LCL}{a vector of lower bound}

  \item{UCL}{a vector of upwer bound}

  \item{...}{further arguments to be passed to
  \code{polygon}}
}
\value{
  a polygon plot
}
\description{
  \code{conf.reg} uses polygon to draw a confidence band
  plot
}

