\name{Dqhat.Ind}
\alias{Dqhat.Ind}
\title{iNterpolation and EXTrapolation of abundance-based Hill number}
\usage{
  Dqhat.Ind(x, q, m)
}
\arguments{
  \item{x}{a vector of species abundances}

  \item{q}{a numerical value of the order of Hill number}

  \item{m}{a integer vector of rarefaction/extrapolation
  sample size}
}
\value{
  a vector of estimated interpolation and extrapolation
  function of Hill number with order q
}
\description{
  \code{Dqhat.Ind} Estimation of interpolation and
  extrapolation of abundance-based Hill number with order q
}

