% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXT.r
\name{ggAsyD}
\alias{ggAsyD}
\title{ggplot for Asymptotic diversity}
\usage{
ggAsyD(outcome)
}
\arguments{
\item{outcome}{the outcome of the functions \code{AsyD} .\cr}
}
\value{
a figure of estimated sample completeness with order q\cr\cr
}
\description{
\code{ggAsyD} Plots q-profile based on the outcome of \code{AsyD} using the ggplot2 package.\cr
It will only show the confidence interval of 'Estimated'.
}
\examples{
## Type (1) example for abundance-based data
## Ex.1
data(spider)
out1 <- AsyD(spider, datatype = "abundance")
ggAsyD(out1)

## Type (2) example for incidence-based data

## Not run:
## Ex.2
data(ant)
out2 <- AsyD(ant, datatype = "incidence_freq")
ggAsyD(out2)

## End(Not run)
}
