\name{iNEXT-package}
\alias{iNEXT-package}
\docType{package}
\title{
iNterpolation and EXTrapolation for species diversity
}
\description{
iNterpolation and EXTrapolation for species diversity: a unified approach for both individual-based (abundance) data and sample-based (incidence) data
}
\details{
\tabular{ll}{
Package: \tab iNEXT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-30\cr
License: \tab GPL (>= 3)\cr
}

}
\author{
T. C. Hsieh [aut, cre, cph]\cr
K. H. Ma [ctb]\cr
Anne Chao [ths]\cr
Maintainer: T. C. Hsieh <euler96@gmail.com>
}
\references{
Chao, A., N. J. Gotelli, T. C. Hsieh, E. L. Sander, K. H. Ma, R. K. Colwell, and A. M. Ellison. (2013). Rarefaction and extrapolation with Hill numbers: a unified framework for sampling and estimation in biodiversity studies, Ecological Monographs \cr
Chao, A., and L. Jost. (2012). Coverage-based rarefaction and extrapolation: standardizing samples by completeness rather than size. Ecology 93:2533-2547. \cr
Colwell, R. K., A. Chao, N. J. Gotelli, S. Y. Lin, C. X. Mao, R. L. Chazdon, and J. T. Longino. (2012). Models and estimators linking individual-based and sample-based rarefaction, extrapolation and comparison of assemblages. Journal of Plant Ecology 5:3-21. \cr
Hsieh, T. C., K. H. Ma, and A. Chao. (2013). iNEXT online: interpolation and extrapolation (Version 1.0) [Software]. Available from http://chao.stat.nthu.edu.tw/blog/software-download/.
}
