% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iNEXT.r
\name{iNEXT}
\alias{iNEXT}
\title{iNterpolation and EXTrapolation of Hill number}
\usage{
iNEXT(x, q = 0, datatype = "abundance", size = NULL, endpoint = NULL,
  knots = 40, se = TRUE, nboot = 50)
}
\arguments{
\item{x}{a vector of species abundances or incidence frequencies. If \code{datatype = "incidence"}, then the input format of first entry must be total number of sampling units, followed by species incidence frequencies.
Note that in iNEXT version 2.0, the input arguments allow the class of \code{\link{numeric}}, \code{\link{matrix}}, \code{\link{data.frame}} (species by sites), or \code{\link{list}}.}

\item{q}{a numeric value, the diversity order of Hill number .}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or
sampling-unit-based incidence data (\code{datatype = "incidence"}).}

\item{size}{an integer vector of sample size (number of individuals or sampling units to compute rarefaction/extrapolation),
default is NULL. If \code{size} is not specified, then rarefaction/extrapolation by the endpoint and knots will be computed.}

\item{endpoint}{an integer of sample size that is the endpoint for rarefaction/extrapolation computation.
Default is double reference sample size.}

\item{knots}{number of equally-spaced knots (say K, default is 40) in the range of sample size between 1 and \code{endpoint}.
If \code{endpoint} is smaller than reference sample size, then \code{iNEXT()} computes rarefaction part only with approximately K equally spaced sizes.
If \code{endpoint} is larger than reference sample size, then \code{iNEXT()} computes approximately K/2 equally spaced sizes for rarefaction part and approximately K/2 equally spaced sizes between reference sample size and the \code{endpoint}.}

\item{se}{calculates the bootstrap standard error and 95\% confidence interval; default is TRUE.}

\item{nboot}{the number of replications, default is 50.}
}
\value{
a list of three objects: \code{DataInfo}, \code{iNextEst} and \code{AsyEst}
}
\description{
\code{iNEXT}: Interpolation and extrapolation of Hill number with order q
}
\examples{
data(spider)
iNEXT(spider, q=0, datatype="abundance")

data(ant)
iNEXT(ant$h500m, q=1, datatype="incidence", size=round(seq(10, 500, length.out=20)), se=FALSE)
}

