% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{print.iNEXT}
\alias{print.iNEXT}
\title{Printing iNEXT object}
\usage{
\method{print}{iNEXT}(x, ...)
}
\arguments{
\item{x}{an \code{iNEXT} object computed by \code{\link{iNEXT}}.}

\item{...}{additional arguments.}
}
\description{
\code{print.iNEXT}: Print method for objects inheriting from class "iNEXT"
}
