% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNEXT_past.R
\name{print.iNEXT_}
\alias{print.iNEXT_}
\title{Printing iNEXT_ object}
\usage{
\method{print}{iNEXT_}(x, ...)
}
\arguments{
\item{x}{an \code{iNEXT_} object computed by \code{\link{iNEXT_}}.}

\item{...}{additional arguments.}
}
\description{
\code{print.iNEXT_}: Print method for objects inheriting from class "iNEXT_"
}
