\name{spider}
\alias{spider}
\docType{data}
\title{
Spider species abundance frequency counts
}
\description{
Spider species abundance data in two canopy manipulation treatments (Ellison et al. 2010, Sackett et al. 2011).
}
\usage{data(spider)}
\format{
  The format is:\cr
List of 2\cr
 $ Girdled: num [1:26] 46 22 17 15 15 9 8 6 6 4 ...\cr
 $ Logged : num [1:37] 88 22 16 15 13 10 8 8 7 7 ...\cr
}
\references{
Ellison, A. M., A. A. Barker-Plotkin, D. R. Foster, and D. A. Orwig. (2010). Experimentally testing the role of foundation species in forests: the Harvard Forest Hemlock Removal Experiment. Methods in Ecology and Evolution 1:168-179.\cr
Sackett, T. E., S. Record, S. Bewick, B. Baiser, N. J. Sanders, and A. M. Ellison. (2011). Response of macroarthropod assemblages to the loss of hemlock (Tsuga canadensis), a foundation species. Ecosphere 2: art74.
}

\keyword{datasets}
