\name{ant}
\alias{ant}
\docType{data}
\title{
Ant data
}
\description{
Species incidence frequency counts for ant samples from five elevations in northeastern Costa Rica (Longino and Colwell 2011)
}
\usage{data(ant)}
\format{
  The first entry is the total number of sampling units, and followed by the incidences abundances. The format is: \cr
List of 5 \cr
 $ h50m  : num [1:228] 599 1 1 1 1 1 1 1 1 1 ... \cr
 $ h500m : num [1:242] 230 1 1 1 1 1 1 1 1 1 ... \cr
 $ h1070m: num [1:123] 150 1 1 1 1 1 1 1 1 1 ... \cr
 $ h1500m: num [1:57] 200 1 1 1 1 1 1 1 1 1 ... \cr
 $ h2000m: num [1:15] 200 1 2 2 3 4 8 8 13 15 ... \cr
 }
\references{
Longino, J. T., and R. K. Colwell. (2011). Density compensation, species composition, and richness of ants on a neotropical elevational gradient. Ecosphere 2:art29.
}

\keyword{datasets}
