% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invChat.R
\name{Auxiliary}
\alias{Auxiliary}
\title{Compute species diversity with fixed sample coverage}
\usage{
Auxiliary(x, datatype = "abundance", base = "size", ref = NULL)
}
\arguments{
\item{x}{a \code{data.frame} or \code{list} for species abundance or incidence frequencies.}

\item{datatype}{the data type of input data. That is individual-based abundance data (\code{datatype = "abundance"}) or sample-based incidence data (\code{datatype = "incidence"}).}

\item{base}{the compared base, That is sample-size-based (\code{base="size"}) or coverage-based (\code{base="coverage"}).}

\item{ref}{the compared reference sample size, sampling units or sample coverage.
If \code{base="size"} and \code{red=NULL}, this function compute the minimum level of sample size or samplping units for all sites.
If \code{base="coverage"} and \code{ref=NULL}, this function compute the minimum level of sample coverage.}
}
\value{
a \code{data.frame} of species diversity table with fixed minimum reference sample size or minimum sample coverage for all sites.
}
\description{
\code{Auxiliary} compute species diversity with fixed sample size or sample coverage.
}
\examples{
data(spider)
Auxiliary(spider, "abundance", base="size")
Auxiliary(spider, "abundance", base="coverage")

data(ant)
Auxiliary(ant, "incidence", base="coverage", ref=0.985)
}

