% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DataInfo}
\alias{DataInfo}
\title{Exhibit basic data information}
\usage{
DataInfo(x, datatype = "abundance")
}
\arguments{
\item{x}{a vector/matrix/list of species frequence}

\item{datatype}{the data type of input data. That is individual-based abundance data (\code{datatype = "abundance"}) or sample-based incidence data (\code{datatype = "incidence"}).}
}
\value{
a data.frame of basic data information
}
\description{
\code{DataInfo} Exhibit basic data information
}
\examples{
data(spider)
DataInfo(spider, datatype="abundance")
}
\author{
T.C. Hsieh
}

