\name{iNEXT-package}
\alias{iNEXT-package}
\docType{package}
\title{
iNterpolation and EXTrapolation for species diversity
}
\description{
iNterpolation and EXTrapolation for species diversity: a unified approach to rarefaction and extrapolation of Hill numbers for individual-based (abundance) data or sampling-unit-based (incidence) data
}
\details{
\tabular{ll}{
Package: \tab iNEXT\cr
Type: \tab Package\cr
Version: \tab 2.0.5\cr
Date: \tab 2015-08-11\cr
License: \tab GPL (>= 3)\cr
}

}
\author{
T. C. Hsieh [aut, cre, cph]\cr
K. H. Ma [ctb]\cr
Anne Chao [ths]\cr
Maintainer: T. C. Hsieh <euler96@gmail.com>
}
\references{
Chao, A., Gotelli, N.J., Hsieh, T.C., Sander, E.L., Ma, K.H., Colwell, R.K. & Ellison, A.M. (2014) Rarefaction and extrapolation with Hill numbers: a framework for sampling and estimation in species diversity studies. Ecological Monographs, 84, 45-67. \cr\cr
Chao, A. & Jost, L. (2012) Coverage-based rarefaction and extrapolation: standardizing samples by completeness rather than size. Ecology, 93, 2533-2547.\cr\cr
Colwell, R.K., Chao, A., Gotelli, N.J., Lin, S.-Y., Mao, C.X., Chazdon, R.L. & Longino, J.T. (2012) Models and estimators linking individual-based and sample-based rarefaction, extrapolation and comparison of assemblages. Journal of Plant Ecology, 5, 3-21. \cr\cr
Hsieh, T.C., Ma, K.H. & Chao, A. (2013) iNEXT online: interpolation and extrapolation (Version 1.0) [Software]. Available from http://chao.stat.nthu.edu.tw/blog/software-download/.
}
