\name{spider}
\alias{spider}
\docType{data}
\title{
Spider data (datatype = "abundance") 
}
\description{
The data include spider species abundances in two canopy manipulation treatments (Girdled and Logged) of hemlock trees (Ellison et al. 2010, Sackett et al. 2011). In the Girdled treatment site, there were 26 species among 168 individuals; in the Logged treatment site, there were 37 species among 252 individuals.
}
\usage{data(spider)}
\format{
  The format for each site is a list of species abundances:\cr
A list of 2 vectors\cr
 $ Girdled: num [1:26] 46 22 17 15 15 9 8 6 6 4 ...\cr
 $ Logged : num [1:37] 88 22 16 15 13 10 8 8 7 7 ...\cr
}
\references{
Ellison, A.M., Barker-Plotkin, A.A., Foster, D.R. & Orwig, D.A. (2010) Experimentally testing the role of foundation species in forests: the Harvard forest hemlock removal experiment. Methods in Ecology and Evolution, 1, 168-179.\cr\cr
Sackett, T.E., Record, S., Bewick, S., Baiser, B., Sanders, N.J. & Ellison, A.M. (2011) Response of macroarthropod assemblages to the loss of hemlock (Tsuga canadensis), a foundation species. Ecosphere, 2, art74.
}

\keyword{datasets}
